+++ src/elections/voter/MaxSingleQualityVoter.java
@@ -16,7 +16,7 @@ public class MaxSingleQualityVoter extends SingleQualityVoter {
     }
 
     @Override
-    public double assess(Candidate candidate) {
+    public int assess(Candidate candidate) {
         return 0;
     }
